backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.3

# Validation & Utils
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
